+++ src/main.js
@@ -1,14 +1,5 @@
 import * as Vue from "vue";
-import * as VueRouter from "vue-router";
+import router from "./routes";
 import App from "./App.vue";
-
-const routes = [
-  // TODO
-  { path: "/", component: App },
-];
-const router = VueRouter.createRouter({
-  history: VueRouter.createWebHistory(),
-  routes,
-});
 
 Vue.createApp(App).use(router).mount("#app");
